refactor

+++ Renderer/SSSharedBufferCache.cpp
@@ -55,21 +55,30 @@ void SSSharedBufferCache::InitializeCube()
 	mCubeIndexBuffer = SSDX11Renderer::GetDX11Device()->CreateIndexBuffer(IndexData);
 		
 	std::vector<VT_PositionNormalTexcoordTangent>& VertexData = SSSharedRenderData::Get().GetCubeVertexData();
-	mCubeVertexBuffer = SSDX11Renderer::GetDX11Device()->CreateVertexBuffer(sizeof(VT_PositionNormalTexcoordTangent), VertexData.size(), VertexData.data());
+	mCubeVertexBuffer = SSDX11Renderer::GetDX11Device()->CreateVertexBuffer(
+		sizeof(VT_PositionNormalTexcoordTangent), 
+		static_cast<unsigned int>(VertexData.size()), 
+		VertexData.data());
 }
 
 
 void SSSharedBufferCache::InitializeSphere()
 {
 	std::vector<VT_PositionNormalTexcoordTangent>& VertexData = SSSharedRenderData::Get().GetSphereVertexData();
-	mSphereVertexBuffer = SSDX11Renderer::GetDX11Device()->CreateVertexBuffer(sizeof(VT_PositionNormalTexcoordTangent), VertexData.size(), VertexData.data());
+	mSphereVertexBuffer = SSDX11Renderer::GetDX11Device()->CreateVertexBuffer(
+		sizeof(VT_PositionNormalTexcoordTangent), 
+		static_cast<unsigned int>(VertexData.size()),
+		VertexData.data());
 }
 
 
 void SSSharedBufferCache::InitializeScreenBlit()
 {
 	std::vector<VT_PositionTexcoord>& VertexData = SSSharedRenderData::Get().GetScreenBlitVertexData();
-	mScreenBlitVertexBuffer = SSDX11Renderer::GetDX11Device()->CreateVertexBuffer(sizeof(VT_PositionTexcoord), VertexData.size(), VertexData.data());
+	mScreenBlitVertexBuffer = SSDX11Renderer::GetDX11Device()->CreateVertexBuffer(
+		sizeof(VT_PositionTexcoord), 
+		static_cast<unsigned int>(VertexData.size()),
+		VertexData.data());
 
 	std::vector<UINT>& IndexData = SSSharedRenderData::Get().GetScreenBlitIndexData();
 	mScreenBlitIndexBuffer = SSDX11Renderer::GetDX11Device()->CreateIndexBuffer(IndexData);
